BUG: don't fail hard when parsing version in GnuFCompiler
When the version could not be determined `None` is supposed to be returned.
However `gnu_version_match` raises an error which leads to hard failures when a "wrong" Fortran compiler is found/used instead of gracefully disabling it which would be completely fine when a Fortran compiler is never used/required.